asynchronous/embassy: Introduce separate interrupt receiver struct
Create a struct to allow waiting for interrupts without needing to block
the core controller object. Move interrupt related code into its own
module.
